Data-Streaming/requirements.txt
kafka-python==1.4.6
confluent_kafka
requests
avro
faust
fastavro
boto3==1.21.19
toml
python-decouple